DB-Project1 v1 complete
I have completed the project according to the project guidelines. I may
be able to improve this project further by taking advantage of Hashmaps
and keeping the database data as String until it is exported into the
database.

edit: updated the teams_by_city operation to support searching cities
with two names.